+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.3.4 (Upcoming)
+
+**Added:**
+- Compute Service - Added support for custom OS disk size during VM creation
+
+**Removed:**
+- Recovery Vault Support
+
 ## 0.3.3
 
 **Added:**